cs493/hw7-chanall/requirements.txt
Flask==2.2.2
google-cloud-datastore==2.9.0
google-auth
requests_oauthlib
python-jose
flask-cors
six
python-dotenv
requests
authlib